backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.0
Pillow==10.4.0
boto3==1.35.0
python-dotenv==1.0.1
pydantic==2.9.0
numpy==1.26.4
